Safari-611.1.10/LayoutTests/fast/forms/change-inputmode-crash-expected.txt

This test verifies that changing the inputmode attribute after changing focus does not cause a crash. To manually run the test, load the page and verify that a crash does not occur.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.activeElement is target
PASS successfullyParsed is true

TEST COMPLETE
